requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.1
pydantic>=2.7.0
rich>=13.7.0
pytest>=8.0.0
pytest-mock>=3.12.0
